requirements.txt
anthropic>=0.30.0
google-genai>=1.0.0
moviepy==1.0.3
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
